move Szumowski to utils and rename; add citations for Kessler '69, Morisson&Grabowski 2007, Arabas et al. 2015
