+++ docs/bibliography.json
@@ -1,7 +1,8 @@
 {
   "https://doi.org/10.1007/978-1-935704-36-2_1": {
     "usages": [
-      "examples/docs/pysdm_examples_landing.md"
+      "examples/docs/pysdm_examples_landing.md",
+      "examples/PySDM_examples/utils/simulation_2D_prescribed_flow/__init__.py"
     ],
     "label": "Kessler 1969 (Meteorol. Monogr. 10)",
     "title": "On the Distribution and Continuity of Water Substance in Atmospheric Circulations"
@@ -89,7 +90,8 @@
       "examples/docs/pysdm_examples_landing.md",
       "examples/PySDM_examples/Morrison_and_Grabowski_2007/fig_2.ipynb",
       "examples/PySDM_examples/Morrison_and_Grabowski_2007/fig_3.ipynb",
-      "examples/PySDM_examples/Morrison_and_Grabowski_2007/__init__.py"
+      "examples/PySDM_examples/Morrison_and_Grabowski_2007/__init__.py",
+      "examples/PySDM_examples/utils/simulation_2D_prescribed_flow/__init__.py"
     ],
     "label": "Morrison & Grabowski 2007 (J. Atmos. Sci. 64)",
     "title": "Comparison of Bulk and Bin Warm-Rain Microphysics Models Using a Kinematic Framework"
@@ -255,7 +257,8 @@
       "examples/docs/pysdm_examples_landing.md",
       "examples/PySDM_examples/Arabas_et_al_2015/gui.ipynb",
       "PySDM/dynamics/displacement.py",
-      "PySDM/physics/particle_advection/implicit_in_space.py"
+      "PySDM/physics/particle_advection/implicit_in_space.py",
+      "examples/PySDM_examples/utils/simulation_2D_prescribed_flow/__init__.py"
     ],
     "label": "Arabas et al. 2015 (Geosci. Model Dev. 2015)",
     "title": "libcloudph++ 1.0: a single-moment bulk, double-moment bulk, and particle-based warm-rain microphysics library in C++"
