+++ examples/PySDM_examples/Arabas_et_al_2015/example.py
@@ -1,6 +1,6 @@
 from open_atmos_jupyter_utils import TemporaryFile
 from PySDM_examples.Arabas_et_al_2015 import Settings, SpinUp
-from PySDM_examples.Szumowski_et_al_1998 import Simulation, Storage
+from PySDM_examples.utils.simulation_2D_prescribed_flow import Simulation, Storage
 from PySDM_examples.utils import DummyController
 
 from PySDM import Formulae
