--- examples/PySDM_examples/Szumowski_et_al_1998/__init__.py
@@ -1,9 +1,0 @@
-# pylint: disable=invalid-name
-"""
-2D GUI-equipped prescribed-flow warm-rain case based on
-[Szumowski et al. 1998 (Atmos. Res.)](https://doi.org/10.1016/S0169-8095(97)00082-3)
-"""
-from .gui_settings import GUISettings
-from .mpdata_2d import MPDATA_2D
-from .simulation import Simulation
-from .storage import Storage
