+++ examples/PySDM_examples/utils/simulation_2D_prescribed_flow/fields.py
