+++ examples/PySDM_examples/utils/simulation_2D_prescribed_flow/make_default_product_collection.py
