+++ tests/smoke_tests/kinematic_2d/arabas_et_al_2015/test_environment.py
@@ -1,7 +1,7 @@
 # pylint: disable=missing-module-docstring,missing-class-docstring,missing-function-docstring
 import numpy as np
 from PySDM_examples.Arabas_et_al_2015 import Settings, SpinUp
-from PySDM_examples.Szumowski_et_al_1998 import Simulation
+from PySDM_examples.utils.simulation_2D_prescribed_flow import Simulation
 
 from PySDM.physics.constants import si
